Bot/requirements.txt
requests==2.32.3
feedparser==6.0.11
readability-lxml==0.8.1
beautifulsoup4==4.12.3
lxml==5.2.2
langdetect==1.0.9
geopy==2.4.1
datasketch==1.6.4
folium==0.17.0
spacy==3.7.5
lxml_html_clean==0.2.0
